Added logic for pattern query answering with variant time but no quantifiers. Also added some unit tests

+++ src/main/java/gr/imsi/athenarc/visual/middleware/patterncache/query/TimeFilter.java
@@ -3,10 +3,10 @@ package gr.imsi.athenarc.visual.middleware.patterncache.query;
 public class TimeFilter {
     // -- Time constraints --
     private final boolean timeAny;    // if true, ignore time constraints
-    private final double timeLow;     // if timeAny == false, min allowable length
-    private final double timeHigh;    // if timeAny == false, max allowable length
+    private final int timeLow;     // if timeAny == false, min allowable length
+    private final int timeHigh;    // if timeAny == false, max allowable length
 
-    public TimeFilter(boolean timeAny, double timeLow, double timeHigh) {
+    public TimeFilter(boolean timeAny, int timeLow, int timeHigh) {
         this.timeAny = timeAny;
         this.timeLow = timeLow;
         this.timeHigh = timeHigh;
@@ -16,11 +16,11 @@ public class TimeFilter {
         return timeAny;
     }
 
-    public double getTimeLow() {
+    public int getTimeLow() {
         return timeLow;
     }
     
-    public double getTimeHigh() {
+    public int getTimeHigh() {
         return timeHigh;
     }
 }
